Add exp. Do not allow nil signatures. Check arguments.
This is a big commit with different parts.

The main one is adding the exponent of a multivector, (exp a).

We also do not allow nil signatures (it used to mean that
all the basis vectors squared to +1).

We check the arguments in all the operations, so they make sense
with numbers and multivectors in general, but no more. It used
to be that (+ "" =) and such things would work and produce 0,
which of course makes no sense.

Generalize (grade a) for numbers too.

Define projection in its general form (with lcontract instead of
prod). Works for degenerate algebras too.

And some more cosmetic changes.