Use dual to define the antiwedge product.
This is probably wrong, but my previous definition was probably wrong too :)

The test passes in any case, and this definition at least makes
the antiwedge work in degenerate algebras (algebras with i*i=0).

It also worked for some examples I compared with from Leo Dorst
and Steven De Keninck.