Fix expressions like -1 in the calculator.
The way we expanded over the operators converted "-1" into "- 1",
which were two different tokens, and it would not understand them as a
single number. Also, the parser would fail, as it expected a number
and not an operator.

This simple change considers - (and +) as functions too, so the parser
applies them even if they have been expanded, and the negative numbers
(or vectors, etc) get recovered correctly.