Many small changes: help shows all documentation, better assignments, etc.
With the environment map mapping to vars instead of functions, we can
easily access now the documentation for every function and operator we
use.

Also, we can exit the calculator by writing "exit" or "quit".

The last result is automatically kept in a variable called "ans" (as
in, the last answer we got, like in octave).

And in general quite a bit of cleanup and reordering for clarity.